+++ frontend/appflowy_flutter/lib/workspace/presentation/settings/widgets/settings_appearance/settings_appearance.dart
@@ -1,0 +1,3 @@
+export 'brightness_setting.dart';
+export 'font_family_setting.dart';
+export 'color_scheme.dart';
